+++ CHANGELOG.md
@@ -6,6 +6,8 @@ backwards compatibility.
 
 ## Unreleased
 
+## 0.2.0
+
 * Introduce signal buffer abstraction.
 * Abstract cells used to access internal audio output and control input data.
 * Implement read-only UI for 128x64 OLED displays.